Adding supplier model and integrating with spree orders and shipments.
This adds a supplier field to line item, product and shipment. When a line item
is created the supplier is copied from the product. When shipments are created
for each supplier, the supplier is taken from the line item.

Also some changes to the checkout flow, removing the delivery step as we automate
the shipping calculation and there is nothing for the user to select.

Signed-off-by: Chris Boertien <chris.boertien@gmail.com>