Overriding add_variant in Order and some cleanup.
add_variant will look for a shipment that matches the supplier of the product being added, if found
the line item is added to the shipment, if not the shipment is created. this will only occur if the
order has a ship_address associated, which will only happen if they have already been to the checkout,
or more importantly, once the user model has a ship_address attached and the user is logged in.

Also fixes a bug in Shipment

Signed-off-by: Chris Boertien <chris.boertien@gmail.com>